pyproject.toml
[tool.isort]
line_length = 100
multi_line_output = 3
profile = "black"

[tool.black]
line-length = 100

[tool.taskipy.tasks]
deploy = "sam validate && poetry export -f requirements.txt --without-hashes > src/requirements.txt && sam build && sam deploy"
format = "echo 'isort:' && isort . && echo 'black:' && black . && echo 'flake8:' && flake8"
lint = "task format"
release = "task lint && task test"
test = "pytest"

[tool.poetry]
name = "cl"
version = "0.1.0"
description = "Craigslist scraper."
authors = ["Tim Coutinho <tmcoutinho42@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
awscli = "^1.18"
beautifulsoup4 = "^4.9"
boto3 = "^1.14"
requests = "^2.23"

[tool.poetry.dev-dependencies]
"aws-cdk.aws-dynamodb" = "^1.85.0"
"aws-cdk.aws-events" = "^1.85.0"
"aws-cdk.aws-events-targets" = "^1.85.0"
"aws-cdk.aws-iam" = "^1.85.0"
"aws-cdk.aws-lambda" = "^1.85.0"
"aws-cdk.core" = "^1.85.0"
aws-sam-cli = "^1.15"
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.7.0"
pytest = "^6.2.1"
taskipy = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
